+++ src/pages/Explore/components/Map/components/LegendControl/LayerOptions.tsx
@@ -39,6 +39,7 @@ export const OpacityCmdButton: React.FC<LayerOptionsProps> = ({
         title="Adjust layer opacity"
         aria-label={`Set ${layer.collection?.title} ${layer.renderOption?.name} opacity`}
         iconProps={{ iconName: "CircleHalfFull" }}
+        role="menuitem"
         styles={
           layer.layer.opacity === 100 ? cmdButtonStyles : activeCmdButtonStyles
         }
